+++ packages/jazz-browser/CHANGELOG.md
@@ -1,5 +1,30 @@
 # jazz-browser
 
+## 0.7.10
+
+### Patch Changes
+
+-   Updated dependencies
+    -   cojson@0.7.10
+    -   cojson-storage-indexeddb@0.7.10
+    -   jazz-tools@0.7.10
+
+## 0.7.9
+
+### Patch Changes
+
+-   Updated dependencies
+    -   cojson@0.7.9
+    -   cojson-storage-indexeddb@0.7.9
+    -   jazz-tools@0.7.9
+
+## 0.7.8
+
+### Patch Changes
+
+-   Updated dependencies
+    -   jazz-tools@0.7.8
+
 ## 0.7.6
 
 ### Patch Changes